server/requirements.txt
blinker==1.8.1
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
numpy==1.26.4
pillow==10.3.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
